refactor: add api service in flow & solana and remove from core
