+++ flow/src/main/java/com/portto/sdk/flow/Flow.kt
@@ -5,7 +5,6 @@ import androidx.annotation.WorkerThread
 import com.portto.sdk.core.Blockchain
 import com.portto.sdk.core.BloctoSDK
 import com.portto.sdk.core.Chain
-import com.portto.sdk.core.network.service.FlowService
 import com.portto.sdk.flow.method.AuthenticateMethod
 import com.portto.sdk.flow.method.SendTransactionMethod
 import com.portto.sdk.flow.method.SignMessageMethod
@@ -13,9 +12,9 @@ import com.portto.sdk.wallet.BloctoSDKError
 import com.portto.sdk.wallet.flow.AccountProofData
 import com.portto.sdk.wallet.flow.CompositeSignature
 
-val BloctoSDK.flow by lazy { Flow() }
+val BloctoSDK.flow by lazy { Flow(FlowService) }
 
-class Flow : Chain {
+class Flow(private val api: FlowService) : Chain {
 
     override val blockchain: Blockchain = Blockchain.FLOW
 
@@ -81,5 +80,5 @@ class Flow : Chain {
      * @return fee payer address
      */
     @WorkerThread
-    fun getFeePayerAddress(): String = FlowService.getFeePayer().address
+    fun getFeePayerAddress(): String = api.getFeePayer().address
 }
