+++ flow/src/main/java/com/portto/sdk/flow/FlowService.kt
@@ -1,0 +1,11 @@
+package com.portto.sdk.flow
+
+import androidx.annotation.WorkerThread
+import com.portto.sdk.core.network.get
+import com.portto.sdk.flow.model.FeePayerResponse
+
+
+object FlowService {
+    @WorkerThread
+    fun getFeePayer(): FeePayerResponse = get("flow/feePayer")
+}
