+++ flow/src/main/java/com/portto/sdk/flow/model/FeePayerResponse.kt
@@ -1,0 +1,6 @@
+package com.portto.sdk.flow.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class FeePayerResponse(val address: String)
