+++ solana/src/main/java/com/portto/sdk/solana/model/SolanaRawTxRequest.kt
@@ -1,0 +1,12 @@
+package com.portto.sdk.solana.model
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SolanaRawTxRequest(
+    @SerialName("sol_address")
+    val address: String,
+    @SerialName("raw_tx")
+    val rawTx: String
+)
